# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Challenge Binar Chapter 8",
    "description": "Api challenge binar chapter 8",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "host api"
    }
  ],
...
  "paths":{
...
    "/super-user/api/{iduser}":{
      "get": {
        "tags": ["dashboard"],
        "summary": "Get detail user id",
        "operationId": "get data",
        "parameters": [
          {
            "name": "iduser",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "succesfull",
            "content":{
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "successfull get data",
                    "statusCode": 200,
                    "result": {  }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}